add recycler view

+++ app/src/main/java/com/example/atm/data/db/dao/UserDao.java
@@ -19,5 +19,7 @@ public interface UserDao {
 
 
     @Query("SELECT currentBalance FROM User WHERE username = :username")
-    Double getCurrentBalanceByUsername(String username);
+    Long getCurrentBalanceByUsername(String username);
+
+
 }
